medical_app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
python-multipart
python-dotenv
psycopg2-binary
pydantic[email]
pillow
numpy==1.26.4
torch==2.0.1
torchvision==0.15.2
alembic>=1.11.1
httpx
python-jose[cryptography]
passlib==1.7.4
python-json-logger
pydantic_settings
openai
bcrypt==4.0.1
aiofiles
